refactor: locate main application in own package
make possible to use microservice architecture
